Extract file item rendering to its own component
