+++ app/src/ui/history/file-list.tsx
@@ -1,12 +1,8 @@
 import * as React from 'react'
 
 import { CommittedFileChange } from '../../models/status'
-
-import { PathLabel } from '../lib/path-label'
 import { List } from '../lib/list'
-
-import { Octicon, iconForStatus } from '../octicons'
-import { mapStatus } from '../../lib/status'
+import { CommittedFileItem } from './committed-file-item'
 
 interface IFileListProps {
   readonly files: ReadonlyArray<CommittedFileChange>
@@ -26,33 +22,12 @@ export class FileList extends React.Component<IFileListProps> {
   }
 
   private renderFile = (row: number) => {
-    const file = this.props.files[row]
-    const status = file.status
-    const fileStatus = mapStatus(status)
-
-    const listItemPadding = 10 * 2
-    const statusWidth = 16
-    const filePathPadding = 5
-    const availablePathWidth =
-      this.props.availableWidth -
-      listItemPadding -
-      filePathPadding -
-      statusWidth
-
     return (
-      <div className="file" onContextMenu={this.props.onContextMenu}>
-        <PathLabel
-          path={file.path}
-          status={file.status}
-          availableWidth={availablePathWidth}
-        />
-
-        <Octicon
-          symbol={iconForStatus(status)}
-          className={'status status-' + fileStatus.toLowerCase()}
-          title={fileStatus}
-        />
-      </div>
+      <CommittedFileItem
+        file={this.props.files[row]}
+        availableWidth={this.props.availableWidth}
+        onContextMenu={this.props.onContextMenu}
+      />
     )
   }
